Add struct mm_imgdesc and related functions
Creates a new type called struct mm_imgdesc which corresponds to an image
description, ie pointer to data buffer is not provided. This description is
different to mmimage in respect to the pixel type description (use a pixel
format) and to the support of a stride parameter that allows to manipulate
non-continous image buffer.

With this image description comes 3 helper functions:
 - size_t mmimg_pixel_size(unsigned int pixel_format)
 - int mmimg_set_stride(struct mm_imgdesc* img, size_t alignment)
 - void* mmimg_alloc_buffer(const struct mm_imgdesc* img)