Add local interprocess communication facilities
Provides set of API to create server/client mechanism supporting file
passing and datagram based communication. This communication allows to
full duplex communication and immediate detection when one of the
endpoint is closed (normally or because process died). On POSIX, this is
implemented with AF_UNIX socket. On Windows, this is based on local
named pipe.

Closes MMLAB-23