Add thread creation and synchronization facilities
Provides new installed header mmthread.h and facilities to create, join or
detach threads as well as synchronization primitives such as mutex,
condition variable and one time initialization. Mutex and condition variable
support being shared across process if initialized supporting it.

On POSIX system, the implementation is a simple wrapper over pthread.
However the implementation on Win32 is more challenging. It requires
reimplementing mutex and condition to support process shared synchronization
primitives which is achieved with the help of a separate service process to
resolve lock contention.

Closes MMLAB-7