Add file manipulation facilities API
Provide the set of API to manipulation file in a platform independent
way. Some of the API required different implementation on POSIX and
Win32. however most of them are small wrapper on top of function
exported by the standard library of the platform. However the API
introduced differ particularly on 2 aspects:
 - the set the mmlib error state in case of error (origin and type of
   error is reported)
 - all file descriptors created are NOT inheritable by default (this
   avoid fd leaks into child process).

Closes MMLAB-13