Declare exported symbol with MMLIB_API
Prefix all symbols in the exported header with the MMLIB_API. If the
header are used not during the compilation of mmlib dll, the macro set
the dllimport attribute to the function declaration. On target different
than Win32, the macro is expanded to nothing.

This mechanism is needed for effortless compatibility with MSVC.

Refs MMLAB-1

Change-Id: Ied5b4a1b76795d0412cf06ef2319f853c7eabb5d