Add MM_CHECK_FUNC and MM_CHECK_FUNCS conftest
Add MM_CHECK_FUNC and MM_CHECK_FUNCS macros for use on configure.ac.
They allows to check for a particular function is available when
specifying library to link with. ie, they are the same as AC_CHECK_FUNC
and AC_CHECK_FUNCS expecting that they accept an optional set of library
to use for linking.

Refs MMLAB-7