Prevent new error being set when logging an error
When mm_raise_error_vfull() is called and error is reported in the log,
there are always possibilities that function used in logging fails. If
one those function report itself mmlib error state, we would end up with
an error state different from the one supposed to be set with
mm_raise_error_vfull(). The changes in this commit ensure that when
error is being reported, new error raise due to logging will be ignored.