path and argv support UTF-8 in mm_spawn on win32
On windows, when argument array and executable path is passed to
mm_spawn(), they are assumed to be UTF-8 encoded. The path and generated
command line is converted to UTF-16 just before calling
CreateProcessW()... instead of calling CreateProcessA().

This allows mm_spawn() to support UTF-8 arguments on windows (they are
already supported on other platform).

Refs MMLAB-26