Make main() return utf-8 argument array on win32
This makes mmlib export 2 functions that are used to override bit of CRT
initialization on Windows. The net result is that if mmlib is directly linked
(is a direct dependencies) of a executable, the command line to argument string
array will be performed by mmlib ensuring that the argument strings are UTF-8.

Refs MMLAB-26