Remove use gl_VISIBILITY in m4/api-exports.m4
In AC_DEF_API_EXPORT_ATTRS autoconf macro, we stop relying on
gl_VISIBILITY to know whether visibility attribute can be used and we
rely solely on the fact that we use a windows host or not. This has the
net advantage that we do not depends on visibility.m4 which is packaged
with gnulib or gettext. The downside is minimal in practice: those
attributes are recognized by most compiler targetting posix platform.
If later we must support non posix nor win32 platform, we will add at
that time the test for compiler support of the attributes.

Refs MMLAB-1