allow dlopen(NULL)
NULL is used to return a handle for the main program.
mm_dlopen() was design to work on plugins (dynamic libraries) where such
a comportement makes no sense.
This new behavior allows a program to launch a function given its name
(as a string) by looking it up inside its own code.

Change-Id: Ib9fa34480b7d770b07c71ff6140fd641b40c834d