declare mm_ino_t, mm_dev_t, uid_t, gid_t
Those types are necessary to implement properly *stat() functions (or
clarify the intent of the fields). Since ino_t and dev_t can be declared
on win32 and their declaration do not match the need on the platform, we
define new type with mm_ prefix. On the other hand uid_t and gid_t are
declared on windows. Hence we declare them with the same POSIX name.

Change-Id: Icad5c99120ec66bf5cecf666762002de0a23eeed