relax test between stat.ctime and mm_open() time
When running the tests of stat related functions, the creation time
(ctime) field reported by the stat function is tested against the time
reported by the system when mm_open() was called. There is always a
slight chance that the reported time differs. Consequently, we allow a
time difference of 1 second. Also the call time() is performed closer to
call to mm_open().

Change-Id: I4a49ac7affa4802b065accb7bf747de6ca9971a5