Add MMARG_PARSER_NOEXIT flag to mmarg_parser
Declare MMARG_PARSER_NOEXIT flag and make use of it with flags field in
struct mmarg_parser to prevent mmarg_parse() to exit in case of argument
parsing and option validatuib failure. If this flag is set, in case of error
or help, the function will return MMARGPARSE_ERROR (-1) or MMARGPARSE_STOP
(-2).

This also allows the callback to request an early stop by returning the
MMARGPARSE_STOP value instead of simply MMARGPARSE_ERROR.

Change-Id: I28dfc2cb9be7a854a9088bf20ed3e04f31faf1ff