add meson build system
This is meant to coexist with the autotools build system.
This will allow a test period for meson as long as we want it to be.

Note: the minimum meson version required here is 0.47. This means that
debian 10 (current stable version) is the minimum supported version.
Debian 9 packaged the 0.37.1 version and is not supported.

Change-Id: I3e0c6dcce5b401aae40ab05e4328d3f6e9d7b2ba