Ensure win32 error meesage does not contain %
If the error message associated with a win32 error code contains formatting
token, this messes with the error message format sent in
mm_raise_error_vfull(). This results the error message not being written
because of failure of vsprintf(). This commit ensure that the Win32 error
description does not contains any '%' character by replacing them with spaces.

Change-Id: Idc9a7b78c81f89f9b83289e492009fc95302d55f