Don't kill threads when calling mm_execv() on win32
Terminated all threads but current upon mm_execv() on win32 apparently
terminates important threads involved in the synchronization primitives.
In particular, this messes with the wait for child termination (leading
to early return).

Since no reliable selection criterion has been found yet to guide the
threads termination, the code is dropped until a better approach is
found (if any is possible).

Now that threads are not terminated in win32 implementation of
mm_execv(), it is possible that the thread in started in
test_execv_process() receive read/write failure upon fd close. This is
normal and they must not trigger abort().

This should solves the random failure of exec_failure test on Windows.

Change-Id: Ie057d51390e9aebc961154e6ac3dcc5653ffcd9b