mm_spawn() and mm_execv() can now search in PATH
The behavior of mm_spawn() and mm_execv() now search for executable in
folder specified in PATH if, and only if, the executable argument
contains a directory separator.

This commit actually harmonized behavior. Indeed, previously this
behavior was or was not already in place depending on the platform and
function (mm_execv() vs mm_spawn()). Now the behavior is clarified.

Change-Id: Ib9710a4c8af873a6811fc0fc6d7151bf3dc3fea5