Modify mmlib fd info passing to child on Win32
On Windows mmlib information regarding opened file descriptor used to be
done through memory map inherited in child. The is now done by using an
extra array added at the end of the CRT buffer passed to child (the
lpReserved2 field of STARTUPINFO).

In addition to be a lighter mechanism, it frees the use of hStd* fields.
Those fields are necessary to remap STDIN, STDOUT or STDERR in a child
that does not use mmlib.

Change-Id: Ia48984b9d805144c5d712e87f4fbc70ea993a92f