Improve support for dir in mm_symlink() in win32
Internally, win32 must track whether a symlink links to a file or a
directory (Necessary to open the directory target of a link). Hence the
type of target must be specified upon symlink creation.

So now when calling mm_symlink() on win32, the type will be detected
from the target if existing and set accordingly in win32 internals. If
target does not exist, the type of symlink is assumed to be link to a
file.

Add test specifically for symlink pointing to dir.

Change-Id: Iff4d509971cf09526f4d765077780b102d4b7bf9