Provide facility for shell completion
When the argument parsing starts to be complex, it becomes easier to
rely on the binary to provide the command line completion directly
instead of reimplementing all the options parsing in a separate shell
script which has often tendency to become outdated when command line
parsing is enriched.

mmarg_parse() now comes with the possibility to provide the shell
completion mechanism if specified by MMARG_PARSER_COMPLETION in flags
field of struct mmarg_parser. In such a case, proposal of completion of
option is listed on stdout if last option. Additionally the API
mmarg_parse_complete() has been added to call the option completion
externally.

Change-Id: Ib57814d6983217fb9de7b7ab354f0df316382083