merge path separator test macros and static inline
IS_PATH_SEPARATOR() is a macro used in src/file.c to test whether a
pointed char is a path separator. mm_dirname() and mm_basename() came
carelessly with their own implementation through static function but on
the char value itself, not pointer.

For consistency within the code base, we need to merge the two in a
common internal header (file-internal.h is a good fit for this). For
various reason (precedence of the pointer dereference operator over
arithmetic), it is easy to misuse a macro that take a pointer to test
the pointed value. Hence we harmonize the code using a static inline
operator on a char value.

Change-Id: I1a045f0051e22813fab16b7620a4b6a536765238