argparse: parser callback support completion state
Change the prototype of callback supplied to mmarg_parse() to provide a
way to indicate whether the parsing is currently completing an option
value. This is useful to provide custom completion logic for an
argument value.

Although this change could have been a ABI break, we will ignore it
because:

 1 - no one is currently using argument parser callback (excepting
 internally in tests and examples of mmlib), hence the break should have
 no impact.

 2 - this is actually not an ABI break on x86_64. The calling
 conventions on it (both MS x64 ABI and System V AMD64 ABI) reserve at
 least 4 registers for the first passed argument and are saved by
 caller. From callee POV, going from 3 to 4 argument just initialize a
 new register which can be ignored given the logic of the callback (we
 can ignore it if completion in not requested in call of mmarg_parse())

Change-Id: I09309be86d49b8f9bcf7bb2140b5dd8f96177bb3