tests: ensure child process use built mmlib
Normally libtool does that for us. However the way it is doing is wrong
and break the information regarded inherited fd passed to the child. On
other platform it is working. What must be done is to keep the mechanism
used by libtool for other platforms and on windows to set PATH to use
the mmlib built locally.

Change-Id: Iee8b98aef064bdb6e9835a47b3618c02c71cd51a