win32: define _WIN32_WINNT globally
For Win32, the macro _WIN32_WINNT specifies the minimal supported
version of the NT kernel. When including the system headers, it will be
used to declare more function than default. It was previously defined at
the top of various source file specific to win32 with different value
defined.

This commit moves the definition from individual C files to global
definition in the build system in src/Makefile.am through AM_CPPFLAGS
defined to the value _WIN32_WINNT_WIN8 (hence we support Windows 8 and
later). This harmonizes the requirement and allows also to avoid to
define it in source file when we move code from a file to another.

Please note that this global definition of _WIN32_WINNT is done ONLY in
src/Makefile.am, ie for compilation of mmlib library and its lock
server. In tests, we set _WIN32_WINNT only for testinternals. It is NOT
set for compiling API tests, ie the tests that use the compiled library
and its exported headers. This is the intended setup: the users of mmlib
do not need to define _WIN32_WINNT to compile code using mmlib. Hence we
need to check whether a mmlib user would have trouble including our
exported headers (tested in api tests) without having the macro defined.

Same has been done for WIN32_LEAN_AND_MEAN which is necessary to avoid
mess when including windows.h.

Change-Id: I5028bde45412de1217c31d757e8431719c96db69