Fix warn about cast between func pointer and void*
Casting pointer to function into pointer to data (even to void*) is
indeed unauthorized in ISO C. A cast through union type is the only
valid cast that can be used. This warning appears on gcc only when
"-pedantic" is passed to compiler flags.

We fix this warning in win32 implementation of mm_dlsym() and in tests
when calling function in child process.

A fix was implemented previously in the tests, but the fix used then
relied on a weird behavior of gcc which was not guaranteed to remain in
future, nor guaranteed to work on any compiler.

Change-Id: I0ad244379a00cc7fc68a6d2819053bb64fabd20b