tests - move to builddir before runing tests
This commit is made to smooth out a difference between the meson and
autotools build systems:
- autotools always executes a binary from the BUILDDIR
  ie. ~/project/build/test/myprog is executed from ~/project/build/test/
- meson executes a binary from the top builddir
  ie ~/project/build/test/myprog is executed from ~/project/build/

This change in the current directory creates difficulties when the test
case involves several binaries at the same time.
Changing the current directory to BUILDDIR removes this difference and
allows us to keep all the tests within the same testapi binary instead
of moving the knowledge of the relative path between the involved binary
to the build system - breaking the current test structure in the
process.

Change-Id: I769c53e69fb0b7d8ddb04bbbe16a748a577d8f69