win32 - add a maximum number of tries to spawn the lockserver
Maximum number of set to 20, which correspond to maximum a second
to try and spawn the server. After which, return an ENOTRECOVERABLE
error.

Also change the logic to spawn a lock server at the start of the test
that will live throughout all the tests instead of spawning a new one
for each tests. On heavily loaded workstations, it can happen that the
process is so slow to spawn that the thread test are trying to connect
to the server before it's up.

This commit should not change anything in the normal cases, but prevent
an infinite loop trying to spawn a server in case the host is stuck in a
bad state and unable to spawn the server for some reason.

Change-Id: I37bdfadfe7688a313e939f0cb55227f4c8bc58bb