Remove def file generation
On windows def file are used to generate the import library compatible
with a given compiler. However generating def file is not supported by
all compiler and this part is a burden to maintain in the build system.

Since it is not difficult to generate a def file from a DLL, we leave
this as post-build process to those interested.

Change-Id: I238a8a4dc993438d2739ac45a2b568c7746fb3ae