api-compat-test tests headers in a temp folder
When executing api-compat-test, header where tested in source directory.
Hence if a public header were including a non install header, it would
pass the test while it shouldn't have.

The test is slightly modified by copying the public headers first in
temporary folder which is used as includedir for the compilation tests.

Change-Id: I4f2e7121eee9992bf7af6273395f895980d95176