Fix reported resolution of MM_CLK_REALTIME on windows
On Windows, GetSystemTimePreciseAsFileTime() provide the clock source of
MM_CLK_REALTIME. Since it returns a 64-bits value representing the
number of 100-nanosecond intervals since January 1, 1601 (UTC), the
resolution of the clock is necessarily 100ns.

Previously this was estimated from the frequency of performance counter
(QPC). While this seems driving the update of realtime timestamp, it is
wrong for various reasons:
- this does not take into account the cropping of 100-ns interval update
- when realtime clock is being adjusted, the reported time do not fall
on multiple QPC interval.

This fixes issue of the clock_resolution unit test reported failure
randomly on windows when testing the case of MM_CLK_REALTIME.

Change-Id: I536ca59278b6535a35153ea0b1e5c177a019f3b1