mm_poll() - ignore negative file descriptors
From the poll man page:
If this [fd] field is negative, then the corresponding events field is ignored
and the revents field returns zero. (This provides an easy way of ignoring a
file descriptor for a single poll() call: simply negate the fd field.

Remove the MM_* wrappers around POLLIN and POLLUP.

Change-Id: I32ab8e4e6b443199604efa415069cb48e5016fb6