mm_mapfile() - clarification on @len parameter
When mapping a file, windows rejects map size greater than the file
size, but linux tolerates, and presents a dummy zeroed memory segment
that may raise a SIGBUG when opened for writing.
POSIX actually rejects the idea of mmaping a file with a size larger
than the mapped object.
Since the whole mmap() idea is performance, let's not interfere with the
Linux implementation. On Windows, change the default (permission) error
into an OVERFLOW error. Update documentation accordingly

Add tests for mm_mapfile() length relative to the mapped file
Do NOT use TEST_SHM because of the teardown callback which locks the
name for use with mm_shm_unlink()

Also, silence logs of shm teardown callback.

Change-Id: I01657bdf20c539a9a91259e76fcf37974c7dfff2