win32: rename file to trash before unlink
File that were meant to be unlinked (removed) were renamed just after
being marked for deletion (and before being actually deleted). They were
renamed to an hidden file in the temp folder or if it failed to same
folder. This operation is meant to make the path available even if the
file could not be removed immediately for various reasons (usually
because the file is still used).

However this renaming is not reliable because:
1 - the file to remove can be on another volume as the temp folder.
2 - the folder of the file to remove could be itself meant to be
removed (like in a recursive folder removal).

To address this issue, the recycle bin folder is used instead: the
recycle bin is a folder present in each mounted volume and a
writeable folder is dedicated in it for each logged user. Hence the
file renaming should never failed as long as the filename used for
renaming is unique. To this end, the file to removed is renamed
following a pattern based on its unique file id on the volume.

On FAT file systems, recycle bin is usually not active. Instead we
move file in the root folder of volume.

Change-Id: I8c59d8b44252a9fc116bfc16ece03cb19ce322ae