win32: mm_unlink() succeed on binaries being used
Previously FILE_FLAG_DELETE_ON_CLOSE was used to mark a file set for deletion.
However, opening file with this flag always fails with access denied error when
opening a executable or dll being loaded.

Instead, the file is now opened simply with DELETE access (not DELETE_ON_CLOSE
flag). This allows to make successfully rename the file and make path
immediately available. Later in the function (instead at the beginning), we
will try to dispose the file (equivalent to add DELETE_ON_CLOSE upon file
opening). This may fail. However, the file which has already been renamed has a
specific filename pattern in the recycle bin path. This pattern will be used to
regularly (every time we push a file to in the recycle bin folder) attempt to
remove the files remaining in the folder.

With this commit, we should now be able to have the same behavior on all
platforms when calling mm_unlink(): the file is actually removed (or will be
soon). No specific usage of the file being remove should prevent it being
unlinked.

Change-Id: Iedeaa06b160d3b7dabc552b3286ed47106a75a87