mmthread - name unions and structs
c99 and C++ do not support anonymous unions and structs. Give them names
in order to make the mmlib API compatible with those standards.

Refactor the previous structs to have two distinct structs for pshared
and swr. integrate the type flag within both, but at the same position
in the structure so that they alias.

Change-Id: I83a7623c186eb714ccf8a40aae0c3372a2d1b0e1