Replace struct timespec with own struct mm_timespec
Instead of using struct timespec for the APIs of mmlib, we use now our
own timespec type: struct mm_timespec.

This structure is meant to be binary compatible with the struct timespec
defined on the platform while avoid the issue of missing declaration of
struct timespec (if compiled with old C standard < C11 without posix
feature) or multiple definition of it (in certain circumstances, on Win32
with mingw using pthread.h).

This is an API break which does not lead to any ABI break.

Change-Id: Ifb34b1ac39b4abe513162f67e22c2471dcfca37d