win32: create lock server pipe before spawn
When a process shared lock is used, the connection to the lockserver is
tested: it try to connect to the named pipe of the lock server. If none
is found, the current process attempt to spawn lock server process and
retry to connect to it. The process is spawn and the first thing the
process try to do is to create the named pipe server end.

However a long time (several dozen of ms) can elapsed between the time a
a client realized that there is no lock server and the time when the
pipe server end is created. During this time other client may have
attempted to create a lock server. And the more processes are created,
the less responsive the system will be and the more chances there are
client creating his own lock server. If the lock is eavily contented,
this may results in a kind of fork bomb.

To sanitize this behavior, the named pipe server is attempted to be
created in the client first and then only the lock server will be
created with the pipe end passed to it. Since only one client will
succeed in creating the name pipe, only one process will be created
which is by far the most expensive part of the lock server
initialization.

Change-Id: I991f559fbede42166138f476e82186ecb0be1bc7