Update patch set 3
Patch Set 3:

> This should allow us to re-enable concurrent mmlib testing on windows, right ?

Not really... First mmlib testing does not currently allows concurrent testing no matter the platform (ipc tests among others prevents this). Second, the reason mmlib is cannot be tested concurrently on win32 is also that if a bug affect the lockserver communication or implementation, this should not affect the result of testing of an unrelated commit (you may run 2 win32 job of 2 different commits in parallel if you do not use the lockable resource)


> Have you tested it ?

I have tested it on mindmazews03 and opensource.mindmaze.com windows builder and I have not noticed issues so far.

Patch-set: 3