Update patch set 1
Patch Set 1: Code-Review+1

This would be fine for me as it is.

However food for thought:
There is no way to signify in mmpack/specs that the version is not current but soon to be. Consequently I feel it is a little too early to bump project version in mmpack/specs (IMO it should have occurred in release commit). But but updating version prevent us to give a proper provides file to the new ABI. The minimal version noted there would be either too low (specifying an incompatible version) either the upcoming version but then mmpack-build would fail because min version of symbols noted in provides would be higher than current version noted in mmpack/specs.

To solve this situation (which would happens in any abi break) the best way, I believe we should do the following:
 - in this commit: Remove libmmlib0.provides ONLY
 - in the release commit (after merge commit): update version in specs add the new libmmlib1.provides

But as I said, the commit is fine for me as is. I've mentioned this only as a seed for dicussion to design proper release processes for later.

Patch-set: 1
Reviewer: Gerrit User 1017133 <1017133@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=+1