test: use different named port on win32 and posix
create_sockclient test uses msnp port to test the decoding of a port not
reserved to privileged processes. Unfortunately, this is no longer
referenced in /etc/services on recent linux distribution. Given the
difficulty to find a named service whose port is > 1024 and known by all
platform, we eventually use a different port name for each platform.

Change-Id: I946f9acf90692a325e8c8b1671fccbc5801e8b14