add default permission flags
Allow mode argument of mm_open(), mm_mkdir() and mm_shm_open() to accept
MODE_DEF or MODE_XDEF optional flags. If set, the permission part is
ignored and default permissions are used.

On POSIX platforms, it corresponds to respectively 0666 and 0777
permissions (before umask to be applied as usual).

On Win32 platforms, it corresponds to use the default security
descriptor, which generally leads to permission inherited from parent
directory.

Change-Id: I39e904cfdbf5521c2978c0f52ad8840f51cae35f