internal_mkdir() handles error state reporting
In order to avoid wrong error state set in case of failed recursive
mkdir, an argument controls whether internal_mkdir() should report error
if path already exists or a component of the path is missing. In such
case, the case is reported as returned value. This return value is used
to only log error in case of recursive creation, or to interpret those
as error in the caller in case of non recursive mm_mkdir().

Change-Id: I852934e060dee5bfc30b017eeeafb980491e3389