win32: fix retval of mm_write() with text file
When writing to text file, the conversion from lf to crlf must be
handled on the fly. When reporting a size that has been written, it must
be in term of lf buffer, and the number of conversion performed on the
written buffer must be removed from the count of written byte. However,
this count was internally tracking the number of byte than must be
written, not those actually written. Hence in case of buffer having more
than 1 conversion, the written number of byte may be underestimated.

This corrects the calculation by using the actual number of byte written
on device.

Change-Id: I1a52b8dd1783a053b10f8f18ae6353d02a5a25cc