Fix testlog failing on linux platforms
setjmp()/longjmp() are used in testlog test to check the proper catch of
of an expected signal, by jumping back to the beginning of the test in
case of signal. However using longjmp() from signal handler is not
portable. While it was working up to some version of linux, now running
this mechanism several time in the life of the process makes subsequent
signal catch to bypass installed signal handler and abort directly.

Using sigsetjmp()/siglongjmp() with saving signal mask corrects the
issue. This should have been used since beginning.

Change-Id: I8b8237241e035f6aca4a18e523387880ad7a9a15