Fix: socket tests fail due to dual protocol services
In the tests if socket API, some services were chosen to elicite an
particular error in test cases because they were believed to be bound to
a single protocol, UDP or TCP. It turns out that some those protocol
have been registered in the IANA registry, maybe in recent revisions, to
both UDP and TCP making those tests fails.

To fix this, new protocols have been chosen: login instead of ssh for
TCP only and biff instead of ntp for UDP only.

Fixes: #7
Change-Id: I316a5f247b4ad4b93626378ca30fd96a06462c25