Allow to disable DNS resquests in tests
Many distributions disallow the use of network access when
building/testing package. Although tests of mmlib do not use much the
network, it is however used in only one place: checking that name
resolution of an inexistent service leads to MM_ENONAME error.

A good approach would be to create a confined environment in which test
are run. However, it is not clear how to achieve this in a cross
platform way (in particular on windows platform).

Hence for the time being, an environment variable is recognized by the
tests, MMLIB_DISABLE_DNSREQ_IN_TESTS, which skip the only test case that
make request to DNS server.

Reference: #7
Change-Id: I782b090c7584274cf0529afabf26796bfb4ef80a